This patch adds the basic support for UnrealIRCD v4.
- As UnrealIRCD v4 now uses SID/UID, commands can now arrive using
the nick OR the uid. To support this the nick_list is now hashed
on the uid and the find_nick() now checks the uid AND the nick.

- The NICK function used to introduce new users has been replaced
by UID, making this version not comptaible with UnrealIRCD < 4.

- Adding support for the UMODE2 command, used when a user changes
his usermodes (ex: becoming IRCop)

- All RAW commands used to set channel modes are now prefixed with
the nick of the bot (the bot will set the mode, not the server).

- Specific version added -unreal4