Support SVID (#15)
* user: Add SVID

We need to track the SVID to be able to use it for Child. With that, we
can then decorrelate the nick from the service account.

* protocol: Use SVID instead of nick to find the user

By doing that, we are decorrelating the user from their account name.
There is still a lot to do to accomodate our commands but this first
attempt should work in most cases.

* protocol: Use SVID in nick_identify

* protocol: Support SVID in m_nick

Before, the svid wasn't checked. This is not an ideal situation yet as
we tied the account name to a nick, it will need to be improved later.

* docker: Bump unrealircd to 5.2.1.1

Also changes around Child's container to allow for better debugeability:
install gdb/start child in gdb.