protocol: Separate Nick from User
With this commit, we are separating nicks from users.  Nick can now be
authenticated as a different User and keep on being logged in when
changing nicks (e.g. Nick->nick != User->nick).

One limitation currently is that one needs to use User->nick as their
nick to be able to identify. Follow up changes will add the option to
specify the User to log into in nick_identify().