net: detach and cleanup remote server on disconnect
Previously, the cleanup_reconnect() function would directly
clear nicks in memory. This would introduce inconsistent state
due to accounts not being de-authed as their corresponding nick
was freed. This led to use-after-free's in subsequent UIDs authed
with such corrupted accounts. The problematic codepath was as
follows:

1. Disconnection event
2. net_disconnect()
3.   cleanup_reconnect()
4.     clear_nicks() -> introducing inconsistent state
5. parse_line()
6.   m_uid()
7.     sasl_finish_session()
8.       killuser() -> killing the nick that is now connecting!
9.   user_login()
10.    NoticeToUser(nptr, ...) -> use-after-free

This commit fixes that by calling detach_server_recursive() on the
connected server, aligning the behaviour when receiving a SQUIT
message. This ensures that nicks are destroyed while keeping
accounts in a consistent state, and also clears the list of
servers from memory.